Create Next Prime Number
So far this program will list out all of the prime numbers for a number hardcoded into the program. Aim is to change this to, firstly, have a user input on this hardcoded number (not hard) and then finally have the user choose when to see the next prime and when to stop.