Pi digits
Most of the  code is there to make it work. It would also, btw, probably work if i just set pi as a value and left it. However i also want to have a function that will calculate pi in there. after ~60 iterations it stops returning a new value of pi, leaving me to suspect there is an overflow error.